refactor: add DebugPanelDialog widget
